Refactoring Functions and Seperating Concerns
Created a function to handle changing the color of
individual pixels. Hooked the various color modes into the
new function, adding compatibility for multiple coloring modes
with arrow controls. I now store the most recent arrow key pressed
so if doubling back on the etch-a-sketch the pixel at the pointer
location is not skipped during coloring.